requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt pyproject.toml
#
markdown-it-py==3.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
pyee==11.1.0
    # via plusdeck (pyproject.toml)
pygments==2.17.2
    # via rich
pyserial==3.5
    # via pyserial-asyncio
pyserial-asyncio==0.6
    # via plusdeck (pyproject.toml)
rich==13.7.0
    # via plusdeck (pyproject.toml)
typing-extensions==4.9.0
    # via pyee
